fix: rename files and change tw class
