+++ src/index.css
@@ -85,7 +85,7 @@
 
   /* Background color for popovers such as <DropdownMenu />, <HoverCard />, <Popover /> */
   --popover: var(--color-crust);
-  --popover-foreground: 222.2 47.4% 11.2%;
+  --popover-foreground: var(--color-text);
 
   /* Default border color */
   --border: 214.3 31.8% 91.4%;
@@ -102,7 +102,7 @@
   --secondary-foreground: 222.2 47.4% 11.2%;
 
   /* Used for accents such as hover effects on <DropdownMenuItem>, <SelectItem>...etc */
-  --accent: var(--color-mantle);
+  --accent: var(--color-crust);
   --accent-foreground: var(--color-text);
 
   /* Used for destructive actions such as <Button variant="destructive"> */
