+++ src/pages/index.ts
@@ -2,9 +2,9 @@ export * from './_404';
 export * from './add-item';
 export * from './create-account';
 export * from './create-list';
-export * from './delete-shopping-list';
 export * from './edit-item';
-export * from './edit-shopping-list';
 export * from './home';
 export * from './login';
-export * from './shopping-list';
+export * from './shopping-list-delete';
+export * from './shopping-list-edit';
+export * from './shopping-list-view';
