+++ src/pages/shopping-list-delete/index.tsx
