+++ src/pages/shopping-list-view/delete-shopping-item.tsx
