+++ src/pages/shopping-list-view/list-item.tsx
