+++ src/pages/shopping-list-view/shopping-list-content.tsx
