+++ src/pages/shopping-list-view/shopping-list-delete-mode.tsx
