+++ src/pages/shopping-list-view/shopping-list-select-mode.tsx
